move editor node/zipper transform functions to editor package
