+++ packages/editor-core/src/parser/__tests__/lexer.test.ts
@@ -1,5 +1,5 @@
-import {row, glyph, frac, subsup, limits} from "../../builders";
-import * as util from "../../util";
+import {row, glyph, frac, subsup, limits} from "../../ast/builders";
+import * as util from "../../ast/util";
 
 import * as Lexer from "../lexer";
 import {serializer} from "../test-util";
