+++ packages/editor-core/src/parser/lexer.ts
@@ -8,7 +8,7 @@
  */
 import {UnreachableCaseError} from "@math-blocks/core";
 
-import * as types from "../types";
+import * as types from "../ast/types";
 
 import {Node, Row, Atom, SourceLocation} from "./types";
