+++ packages/editor-core/src/printer/__tests__/printer.test.ts
@@ -1,8 +1,8 @@
 import * as Semantic from "@math-blocks/semantic";
 
-import * as builders from "../../builders";
-import * as types from "../../types";
-import * as Util from "../../util";
+import * as builders from "../../ast/builders";
+import * as types from "../../ast/types";
+import * as Util from "../../ast/util";
 import {toEqualEditorNode} from "../../test-util";
 
 import {print} from "../printer";
