+++ packages/editor-core/src/reducer/__tests__/convert.test.ts
@@ -1,4 +1,4 @@
-import * as builders from "../../builders";
+import * as builders from "../../ast/builders";
 
 import {zipperToRow} from "../convert";
 import {row, toEqualEditorNodes, zrow} from "../test-util";
