+++ packages/editor-core/src/reducer/__tests__/move-cursor.test.ts
@@ -1,5 +1,5 @@
-import * as types from "../../types";
-import * as builders from "../../builders";
+import * as types from "../../ast/types";
+import * as builders from "../../ast/builders";
 
 import {moveLeft} from "../move-left";
 import {moveRight} from "../move-right";
