+++ packages/editor-core/src/reducer/insert-char.ts
@@ -1,4 +1,4 @@
-import * as builders from "../builders";
+import * as builders from "../ast/builders";
 
 import type {Zipper} from "./types";
