+++ packages/editor-core/src/reducer/test-util.ts
@@ -1,7 +1,7 @@
 import * as Semantic from "@math-blocks/semantic";
 
-import * as types from "../types";
-import * as builders from "../builders";
+import * as types from "../ast/types";
+import * as builders from "../ast/builders";
 
 import type {ZRow} from "./types";
