+++ packages/editor-core/src/reducer/types.ts
@@ -1,4 +1,4 @@
-import * as types from "../types";
+import * as types from "../ast/types";
 
 export type ZRow = {
     id: number;
